convert and refactor
