+++ README.md
@@ -44,7 +44,7 @@ STM32F2            |             |             |     X       |
 Adafruit invests time and resources providing this open source code, please support Adafruit and open-source hardware by purchasing products from Adafruit!
 
 # Dependencies
- * [TinyWireM](https://github.com/adafruit/TinyWireM)
+ * [Adafruit BusIO](https://github.com/adafruit/Adafruit_BusIO)
 
 # Contributing
