+++ library.properties
@@ -1,5 +1,5 @@
 name=RTClib
-version=1.14.2
+version=2.0.0
 author=Adafruit
 maintainer=Adafruit <info@adafruit.com>
 sentence=A fork of Jeelab's fantastic RTC library
@@ -7,4 +7,4 @@ paragraph=Works with DS1307, DS3231, PCF8523, PCF8563 on multiple architectures
 category=Timing
 url=https://github.com/adafruit/RTClib
 architectures=*
-depends=TinyWireM
+depends=Adafruit BusIO
