+++ src/RTC_Millis.cpp
@@ -1,0 +1,27 @@
+#include "RTClib.h"
+
+/**************************************************************************/
+/*!
+    @brief  Set the current date/time of the RTC_Millis clock.
+    @param dt DateTime object with the desired date and time
+*/
+/**************************************************************************/
+void RTC_Millis::adjust(const DateTime &dt) {
+  lastMillis = millis();
+  lastUnix = dt.unixtime();
+}
+
+/**************************************************************************/
+/*!
+    @brief  Return a DateTime object containing the current date/time.
+            Note that computing (millis() - lastMillis) is rollover-safe as long
+            as this method is called at least once every 49.7 days.
+    @return DateTime object containing current time
+*/
+/**************************************************************************/
+DateTime RTC_Millis::now() {
+  uint32_t elapsedSeconds = (millis() - lastMillis) / 1000;
+  lastMillis += elapsedSeconds * 1000;
+  lastUnix += elapsedSeconds;
+  return lastUnix;
+}
